Fix account null
git-svn-id: https://svn.code.sf.net/p/geokretylog/code@200 ff536b33-f402-454f-b109-af6ef765b36a

+++ trunk/src/pl/nkg/geokrety/activities/AccountActivity.java
@@ -345,7 +345,7 @@ public class AccountActivity extends ManagedDialogsActivity implements LocationL
     protected void onRestoreInstanceState(final Bundle savedInstanceState) {
         super.onRestoreInstanceState(savedInstanceState);
         secid = savedInstanceState.getString(User.SECID);
-        accountID = savedInstanceState.getInt(User.ACCOUNT_ID);
+        accountID = savedInstanceState.getLong(User.ACCOUNT_ID);
         ocUUIDs = savedInstanceState.getStringArray(User.OCUUIDS);
         ocLogins = savedInstanceState.getStringArray(User.OCLOGINS);
         accountName = savedInstanceState.getString(User.ACCOUNT_NAME);
